cast to int for * in $.*s

+++ server/dhcp.c
@@ -1087,8 +1087,8 @@ void dhcpinform (packet, ms_nulltp)
 		if (i >= sizeof(raw.file)) {
 			log_info("file name longer than packet field "
 				 "truncated - field: %lu name: %d %.*s", 
-				 (unsigned long)sizeof(raw.file), i, i,
-				 d1.data);
+				 (unsigned long)sizeof(raw.file), i,
+				 (int)i, d1.data);
 			i = sizeof(raw.file);
 		} else
 			raw.file[i] = 0;
@@ -1107,8 +1107,8 @@ void dhcpinform (packet, ms_nulltp)
 		if (i >= sizeof(raw.sname)) {
 			log_info("server name longer than packet field "
 				 "truncated - field: %lu name: %d %.*s", 
-				 (unsigned long)sizeof(raw.sname), i, i,
-				 d1.data);
+				 (unsigned long)sizeof(raw.sname), i,
+				 (int)i, d1.data);
 			i = sizeof(raw.sname);
 		} else
 			raw.sname[i] = 0;
@@ -3022,7 +3022,7 @@ void dhcp_reply (lease)
 			log_info("file name longer than packet field "
 				 "truncated - field: %lu name: %d %.*s", 
 				 (unsigned long)sizeof(raw.file),
-				 state->filename.len, state->filename.len,
+				 state->filename.len, (int)state->filename.len,
 				 state->filename.data);
 	} else
 		bufs |= 1;
@@ -3042,7 +3042,7 @@ void dhcp_reply (lease)
 				 "truncated - field: %lu name: %d %.*s", 
 				 (unsigned long)sizeof(raw.sname),
 				 state->server_name.len,
-				 state->server_name.len,
+				 (int)state->server_name.len,
 				 state->server_name.data);
 	} else
 		bufs |= 2; /* XXX */
